Update map setup for MPU support
For imxrt106x, current FlexRAM setup is 0xaabfffff, that
means dtcm=5*32kB (160kB), itcm: 11*32kB (352kB).
Adding AIPS1-4, 5 (peripherals) and FlexSPI memory maps.
This setup was tested on evk-imxrt1064 and is the minimum required
to enable MPU with `dummyfs`, `imxrt-multi` and `psh`.
The activation of MPU has been tested on imxrt106x target.

For imxrt-117x, adjust the alignment of ocram2 to 512*1024 to make
the pre-setting of MPU regions and memory maps in plo work.
The activation of MPU has not been tested on imxrt117x target.

JIRA: RTOS-2