scripts: ia32-generic: set cpu to supported pentium3
When some ca-certificates are used by openssl, sse2 instructions may be executed.
Openssl firstly checks wheather an extension exists, but when it does and isn't supported in kernel, illegal instructions are executed.
sse2 can't also be disabled (on processors with it, e. g. pentium4), even if it's not supported by the os.
That's why the cpu is set to supported one in ia32-generic scripts - pentium3.

JIRA: PD-251